Rework webhook, WIP
