+++ src/main/java/org/jenkinsci/plugins/ghprb/GhprbGitHubAuth.java
@@ -45,6 +45,9 @@ import hudson.model.Item;
 import hudson.security.ACL;
 import hudson.util.FormValidation;
 import hudson.util.ListBoxModel;
+import javax.crypto.Mac;
+import javax.crypto.spec.SecretKeySpec;
+import org.apache.commons.codec.binary.Hex;
 
 public class GhprbGitHubAuth extends AbstractDescribableImpl<GhprbGitHubAuth> {
     private static final Logger logger = Logger.getLogger(GhprbGitHubAuth.class.getName());
@@ -112,6 +115,41 @@ public class GhprbGitHubAuth extends AbstractDescribableImpl<GhprbGitHubAuth> {
     @Exported
     public String getSecret() {
         return secret;
+    }
+    
+    // Check the signature of the message sent from github against the
+    // secret in this authentication context.
+    public boolean checkSignature(String body, String signature) {
+        String secret = getSecret();
+        if (StringUtils.isEmpty(secret)) {
+            return true;
+        }
+        
+        if (signature != null && signature.startsWith("sha1=")) {
+            String expected = signature.substring(5);
+            String algorithm = "HmacSHA1";
+            try {
+                SecretKeySpec keySpec = new SecretKeySpec(secret.getBytes(), algorithm);
+                Mac mac = Mac.getInstance(algorithm);
+                mac.init(keySpec);
+                byte[] localSignatureBytes = mac.doFinal(body.getBytes("UTF-8"));
+                String localSignature = Hex.encodeHexString(localSignatureBytes);
+                if (! localSignature.equals(expected)) {
+                    logger.log(Level.SEVERE, "Local signature {0} does not match external signature {1}",
+                            new Object[] {localSignature, expected});
+                    return false;
+                }
+            } catch (Exception e) {
+                logger.log(Level.SEVERE, "Couldn't match both signatures");
+                return false;
+            }
+        } else {
+            logger.log(Level.SEVERE, "Request doesn't contain a signature. Check that github has a secret that should be attached to the hook");
+            return false;
+        }
+
+        logger.log(Level.INFO, "Signatures checking OK");
+        return true;
     }
 
     private static GitHubBuilder getBuilder(Item context, String serverAPIUrl, String credentialsId) {
